refactor: extract candidateScorer from queueManipulation (phase 3.1)
